Split location provider and resolver tests #292
